// frontend/src/enum/url-path.enum.ts
export enum UrlPathEnum {
  PATH_HOME = '/',
  PATH_LOGIN = "login",
  PATH_DOCS_SHOW = "docs",
  PATH_DOC_CREATE = "doc/create",
  PATH_DOC_SHOW = "doc/show/:id",
  PATH_404 = '(.*)',
  PATH_DOC_EDIT = "doc/edit/:id",
  PATH_USERS_SHOW = 'users',
  CALL_CENTER = 'callcenter',
  INVENTORY = "inventory",
}